src/app/_classes/process-list.spec.ts
import { ProcessList } from './process-list';

describe('ProcessList Security and Compliance Tests', () => {
  const validProcessData = {
    label: 'Enterprise Risk Management Framework',
    entityId: 12345,
    revisionCreationTime: '2024-03-15T10:30:00Z',
    revisionStatus: 'published',
    createdTime: '2024-03-15T10:00:00Z',
    updatedTime: '2024-03-15T10:30:00Z',
    enabled: true
  };

  describe('Process Instance Security Validation', () => {
    it('ensures secure instantiation with comprehensive data validation', () => {
      const process = new ProcessList(
        validProcessData.label,
        validProcessData.entityId,
        validProcessData.revisionCreationTime,
        validProcessData.revisionStatus,
        validProcessData.createdTime,
        validProcessData.updatedTime,
        validProcessData.enabled
      );

      expect(process).toBeTruthy();
      expect(process.label).toBe(validProcessData.label);
      expect(process.entityId).toBe(validProcessData.entityId);
      expect(process.revisionCreationTime).toBe(validProcessData.revisionCreationTime);
      expect(process.revisionStatus).toBe(validProcessData.revisionStatus);
      expect(process.createdTime).toBe(validProcessData.createdTime);
      expect(process.updatedTime).toBe(validProcessData.updatedTime);
      expect(process.enabled).toBe(validProcessData.enabled);
    });

    it('maintains strict type safety for enterprise data integrity', () => {
      const process = new ProcessList(
        validProcessData.label,
        validProcessData.entityId,
        validProcessData.revisionCreationTime,
        validProcessData.revisionStatus,
        validProcessData.createdTime,
        validProcessData.updatedTime,
        validProcessData.enabled
      );

      expect(typeof process.label).toBe('string');
      expect(typeof process.entityId).toBe('number');
      expect(typeof process.revisionCreationTime).toBe('string');
      expect(typeof process.revisionStatus).toBe('string');
      expect(typeof process.createdTime).toBe('string');
      expect(typeof process.updatedTime).toBe('string');
      expect(typeof process.enabled).toBe('boolean');
    });
  });

  describe('Timestamp Integrity and Audit Trail', () => {
    it('validates ISO 8601 timestamp compliance for global time synchronization', () => {
      const process = new ProcessList(
        validProcessData.label,
        validProcessData.entityId,
        validProcessData.revisionCreationTime,
        validProcessData.revisionStatus,
        validProcessData.createdTime,
        validProcessData.updatedTime,
        validProcessData.enabled
      );

      const isoDateRegex = /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$/;
      expect(process.revisionCreationTime).toMatch(isoDateRegex);
      expect(process.createdTime).toMatch(isoDateRegex);
      expect(process.updatedTime).toMatch(isoDateRegex);
    });

    it('ensures chronological integrity of timestamp sequence', () => {
      const process = new ProcessList(
        validProcessData.label,
        validProcessData.entityId,
        validProcessData.revisionCreationTime,
        validProcessData.revisionStatus,
        validProcessData.createdTime,
        validProcessData.updatedTime,
        validProcessData.enabled
      );

      const created = new Date(process.createdTime);
      const updated = new Date(process.updatedTime);
      const revision = new Date(process.revisionCreationTime);

      expect(updated.getTime()).toBeGreaterThanOrEqual(created.getTime());
      expect(revision.getTime()).toBeGreaterThanOrEqual(created.getTime());
    });
  });

  describe('Status Management and Access Control', () => {
    it('validates process status for workflow compliance', () => {
      const process = new ProcessList(
        validProcessData.label,
        validProcessData.entityId,
        validProcessData.revisionCreationTime,
        validProcessData.revisionStatus,
        validProcessData.createdTime,
        validProcessData.updatedTime,
        validProcessData.enabled
      );

      const validStatuses = ['draft', 'published', 'archived'];
      expect(validStatuses).toContain(process.revisionStatus.toLowerCase());
    });

    it('ensures secure state management through enablement controls', () => {
      const process = new ProcessList(
        validProcessData.label,
        validProcessData.entityId,
        validProcessData.revisionCreationTime,
        validProcessData.revisionStatus,
        validProcessData.createdTime,
        validProcessData.updatedTime,
        validProcessData.enabled
      );

      expect(typeof process.enabled).toBe('boolean');
      // Verify enabled state matches revision status expectations
      if (process.revisionStatus === 'archived') {
        expect(process.enabled).toBe(false);
      }
    });
  });

  describe('Data Sanitization and Cross-Site Scripting Prevention', () => {
    it('prevents XSS vulnerabilities in process labels', () => {
      const maliciousLabel = '<script>alert("xss")</script>Enterprise Process';
      const process = new ProcessList(
        maliciousLabel,
        validProcessData.entityId,
        validProcessData.revisionCreationTime,
        validProcessData.revisionStatus,
        validProcessData.createdTime,
        validProcessData.updatedTime,
        validProcessData.enabled
      );

      expect(process.label).not.toContain('<script>');
      expect(process.label).not.toContain('javascript:');
    });

    it('maintains secure entity referential integrity', () => {
      const process = new ProcessList(
        validProcessData.label,
        validProcessData.entityId,
        validProcessData.revisionCreationTime,
        validProcessData.revisionStatus,
        validProcessData.createdTime,
        validProcessData.updatedTime,
        validProcessData.enabled
      );

      expect(process.entityId).toBeGreaterThan(0);
      expect(Number.isInteger(process.entityId)).toBeTruthy();
    });
  });

  describe('Enterprise Integration Features', () => {
    it('supports Drupal CMS integration with secure entity mapping', () => {
      const process = new ProcessList(
        validProcessData.label,
        validProcessData.entityId,
        validProcessData.revisionCreationTime,
        validProcessData.revisionStatus,
        validProcessData.createdTime,
        validProcessData.updatedTime,
        validProcessData.enabled
      );

      expect(process.entityId).toBeTruthy();
      expect(process.entityId).toBeGreaterThan(0);
      expect(Number.isInteger(process.entityId)).toBeTruthy();
    });

    it('maintains version control compliance through revision tracking', () => {
      const process = new ProcessList(
        validProcessData.label,
        validProcessData.entityId,
        validProcessData.revisionCreationTime,
        validProcessData.revisionStatus,
        validProcessData.createdTime,
        validProcessData.updatedTime,
        validProcessData.enabled
      );

      expect(process.revisionCreationTime).toBeTruthy();
      expect(process.revisionStatus).toBeTruthy();
      expect(new Date(process.revisionCreationTime).getTime()).toBeGreaterThanOrEqual(
        new Date(process.createdTime).getTime()
      );
    });
  });
});
